isaac_lab_training/requirements.txt
gymnasium>=0.28
numpy>=1.23
stable-baselines3>=2.0.0
torch>=2.0
onnx>=1.14
onnxruntime>=1.15
pyyaml>=6.0
pytest>=7.0
